+++ app/src/main/java/com/example/juanperezdealgaba/sac/ConcreteTest.java
@@ -436,7 +436,7 @@ public class ConcreteTest extends AppCompatActivity implements AdapterView.OnIte
 
                 while (System.currentTimeMillis() < startTime + maxDurationInMilliseconds) {
                     mbedTLS test = new mbedTLS();
-                    double[] timesAES = test.AESCBC();
+                    int[] timesAES = test.AESCBC();
 
                     System.out.println("Time to encrypt:" + timesAES[0] + "ns\n");
                     writer.write("Time to encrypt:" + timesAES[0] + "ns\n");
